Add files via upload
Data Modelling---Added Date Table and Fact Table. Made connection manually between the new tables. Created Measures to calculate yearly and quaterly sales. 
We can conclude the following thing :-

(a) From Quarterly graph we can say that from year 2012 to 2015 quarter 4 i.e. Oct to Dec. the sales is high .
(b) Secondly July to Sept. it's low  Sales.
(c) Looking over Yearly graph we observe in 2012 sales was 3.2 M which is low , then in 2013 it raised high to 7.2 M again went low in 2014 to 6.4M and gain very high in 2015 to 15.1M,  it's almost double of 2013 and 2014.